requirements.txt
# --- API Framework ---
fastapi==0.111.1
uvicorn[standard]==0.30.1

# --- Core Data Science & ML (Pinned for stability) ---
imbalanced-learn==0.12.3
joblib==1.4.2
lightgbm==4.4.0
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
seaborn==0.13.2

# --- Deep Learning (Pinned for stability) ---
Pillow==10.4.0
scikit-image==0.24.0
tensorflow-cpu==2.16.2
tf-keras==2.16.0 

# --- Chatbot & RAG (LangChain) ---
google-generativeai
langchain
langchain-community
langchain-google-genai
langchain-huggingface
pinecone-client==3.2.2
langchain-pinecone==0.0.3
pypdf
sentence-transformers
torch
torchvision
tiktoken

# --- Experiment Tracking ---
mlflow==2.14.3

# --- Utilities ---
python-dotenv==1.0.1
tqdm==4.66.4

# --- Compatibility fix ---
protobuf==4.25.3

pytest
